requirements.txt
torch
torchvision
ftfy
regex
tqdm
clip @ git+https://github.com/openai/CLIP.git
faiss-cpu
Flask